+++ common/src/main/java/io/github/mortuusars/exposure/item/PhotographItem.java
@@ -7,18 +7,14 @@ import io.github.mortuusars.exposure.Config;
 import io.github.mortuusars.exposure.Exposure;
 import io.github.mortuusars.exposure.PlatformHelper;
 import io.github.mortuusars.exposure.camera.infrastructure.FrameData;
-import io.github.mortuusars.exposure.entity.PhotographEntity;
 import io.github.mortuusars.exposure.gui.ClientGUI;
 import io.github.mortuusars.exposure.gui.component.PhotographTooltip;
 import io.github.mortuusars.exposure.util.ItemAndStack;
 import net.minecraft.ChatFormatting;
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.StringUtil;
 import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.SlotAccess;
 import net.minecraft.world.entity.player.Player;
@@ -28,9 +24,7 @@ import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.gameevent.GameEvent;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -100,33 +94,6 @@ public class PhotographItem extends Item {
                 }
             }
         }
-    }
-
-    @Override
-    public @NotNull InteractionResult useOn(UseOnContext context) {
-        BlockPos clickedPos = context.getClickedPos();
-        Direction direction = context.getClickedFace();
-        BlockPos resultPos = clickedPos.relative(direction);
-        Player player = context.getPlayer();
-        ItemStack itemStack = context.getItemInHand();
-        if (player == null || player.level().isOutsideBuildHeight(resultPos) || !player.mayUseItemAt(resultPos, direction, itemStack))
-            return InteractionResult.FAIL;
-
-        Level level = context.getLevel();
-        PhotographEntity photographEntity = new PhotographEntity(level, resultPos, direction, itemStack.copy());
-
-        if (photographEntity.survives()) {
-            if (!level.isClientSide) {
-                photographEntity.playPlacementSound();
-                level.gameEvent(player, GameEvent.ENTITY_PLACE, photographEntity.position());
-                level.addFreshEntity(photographEntity);
-            }
-
-            itemStack.shrink(1);
-            return InteractionResult.sidedSuccess(level.isClientSide);
-        }
-
-        return InteractionResult.FAIL;
     }
 
     @Override